+++ data/PetNames.lua
@@ -1,0 +1,52 @@
+--[[
+    LuiExtended
+    License: The MIT License (MIT)
+--]]
+
+-- Quests namespace
+LUIE.Data.PetNames = {}
+local PetNames = LUIE.Data.PetNames
+
+PetNames.Sorcerer = {
+    ["Familiar"] = true,
+    ["Clannfear"] = true,
+    ["Volatile Familiar"] = true,
+    ["Winged Twilight"] = true,
+    ["Twilight Tormentor"] = true,
+    ["Twilight Matriarch"] = true,
+    ["Summoned Storm Atronach"] = true,
+    ["Greater Storm Atronach"] = true,
+    ["Charged Atronach"] = true,
+}
+
+PetNames.Warden = {
+    ["Feral Guardian"] = true,
+    ["Eternal Guardian"] = true,
+    ["Wild Guardian"] = true,
+}
+
+PetNames.Necromancer = {
+    ["Skeletal Mage"] = true,
+    ["Skeletal Archer"] = true,
+    ["Skeletal Arcanist"] = true,
+    ["Spirit Mender"] = true,
+    ["Spirit Guardian"] = true,
+    ["Intensive Mender"] = true,
+
+}
+
+PetNames.Sets = {
+    ["Dwarven Sphere"] = true,
+    ["Daedroth"] = true,
+    ["Shadowrend"] = true,
+    ["Dwemer Spider"] = true,
+    ["Morkuldin Sword"] = true,
+}
+
+PetNames.Assistants = {
+    ["Ezabi"] = true,
+    ["Fezez"] = true,
+    ["Pirharri the Smuggler"] = true,
+    ["Nuzhimeh"] = true,
+    ["Tythis"] = true,
+}